+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.67.6
 
 - Retire old SVR2 staging enclave
+
+- keytrans: Remove unused APIs